requirements.txt
ovos-plugin-manager>=0.0.24
ovos-utils~=0.0,>=0.0.30
SpeechRecognition~=3.8
huggingface_hub==0.23.2
nemo_toolkit[all]